DJS - Did some cosmetic work to Home and New Event
I worked on the home page, the title now displays the email of the user.
The new event activity had some work done with scaling trap counters,
buttons, and text. I'll play around with this a bit more later. Need to
work on the rest of the core functionality. Next step, processing a new
event when saved.