DJS - Adding editing and deleting of load and gun
Added functionality to edit and delete load or gun from the listview in
the armory. I was also able to condense the add gun or load into the
edit function since they are virtually the same dialog. There were a
couple of questionable booleans used that might not be the most
intuitive, but they work! Going to work on integrating Google Drive to
save database there. Might be a wild ride...