DJS - A lot of good happening
I'm working with the custom list adapter and a lot of things are going
right. I've updated the remove and edit dialogs for gun and load to be
apart of the object class. I've updated the list item design and added
functionality to the buttons. I removed unused code from Armory
Activity.