DJS - Redid the trap score widget view
I redid the trap score widget and it looks and works a lot better than
the previous model. I'm pretty happy with it, but it still needs a
little work. It works with new and quick events. I will try to work on
adding functionality for rounds and for teams.