DJS - New Event and Shooter Class is close
I worked on the new event, shooter, and post event classes today. I had
to rework the event and shot classes themselves and the work for them in
the DBHandler class. Need to work on saving shots in post event and how
post event displays. Then need to work with the event history page. Then
this might be pretty close to done for an alpha release.