DJS - Restyled the post event item
I restyled the post event item to look a bit cleaner. We'll see if that
is true or not. Next, I'm going to add rounds to the mix for new events
and post events. I might scrap the quick event feature for right now.
See a bit much to handle with the state of this.