DJS - Updated Database and classes
I updated the DB and classes associated with it to reorganize the data structures used in this app. Hopefully this is more streamlined and will allow for easier reusability. A shooting event will now take the form: Round -> Match -> Event. Where each round has an object, a match is made of rounds, and a match can pair and event to it.